# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
...
        "/api/artista": {
            "get": {
                "tags": [
                    "Artistas"
                ],
                "summary": "Todos os Artistas",
                "description": "Lista todos os Artistas com seus albuns",
                "operationId": "415b1f39080db3f4e460e2418e0f6cfa",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº da página",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Artistas"
                ],
                "summary": "Cadastra novo Artista",
                "description": "Registra um novo Artista",
                "operationId": "cc47ff4349a86cabe02cf497c6d42539",
                "parameters": [
                    {
                        "name": "art_nome",
                        "in": "query",
                        "description": "Nome",
                        "required": true
                    },
                    {
                        "name": "art_descricao",
                        "in": "query",
                        "description": "Descrição",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Artista não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/artista/{id}": {
            "get": {
                "tags": [
                    "Artistas"
                ],
                "summary": "Mostra um Artista",
                "description": "Pesquisa por um artista através do (id)",
                "operationId": "7f15968063fb6e299a3574a3a4aad99b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Nº de identificação do artista",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artista Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Pessoa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Artistas"
                ],
                "summary": "Atualizar dados de um Artista",
                "description": "Editar os dados de um artista através do (id)",
                "operationId": "ad7ddf3d1210e2353559d9a6e7e29d45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "art_nome",
                                    "art_descricao"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "art_nome": {
                                        "type": "string",
                                        "example": "Nome artista"
                                    },
                                    "art_descricao": {
                                        "type": "string",
                                        "example": "Descrição artista"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artista atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Artista atualizado com sucesso"
                                        },
                                        "artista": {
                                            "properties": {
                                                "art_nome": {
                                                    "type": "string",
                                                    "example": "Nome Artista"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "404": {
                        "description": "Endereco não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Artistas"
                ],
                "summary": "Exclui um Artista",
                "description": "Exclui um artista através do (id)",
                "operationId": "d083dded75c0c9de9e0b5b3007c231c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do artista",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "artista excluído com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Não foi possível excluir o album",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/artista/{pesquisa}": {
            "get": {
                "tags": [
                    "Artistas"
                ],
                "summary": "Pesquisar um Artista",
                "description": "Pesquise pelo nome do artista",
                "operationId": "19b43f69ea24019e21b2b5be53259e05",
                "parameters": [
                    {
                        "name": "pesquisa",
                        "in": "path",
                        "description": "Pesquise pelo nome do artista",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Artista Exemplo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artista Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Artista não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/foto-artista": {
            "post": {
                "tags": [
                    "Foto Upload"
                ],
                "summary": "Faz o upload de uma foto de artista",
                "description": "Upload de Arquivo",
                "operationId": "b2259c653746dd8c6a2c15719ea88d8e",
                "parameters": [
                    {
                        "name": "artista_id",
                        "in": "query",
                        "description": "Nº de identificação do Artista",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Imagem do Artista enviado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "artista_id": {
                                            "type": "integer"
                                        },
                                        "file": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro no upload"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}